requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
cached-property==1.5.2
certifi==2024.7.4
cffi==1.17.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
convertdate==2.4.0
cryptography==42.0.8
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
docutils==0.21.2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
exchangelib==5.4.2
fonttools==4.53.1
fpdf2==2.7.5
furl==2.1.3
graphviz==0.13.2
h11==0.14.0
hijri-converter==2.3.1
holidays==0.21.13
htmldocx==0.0.6
idna==3.7
importlib_metadata==8.4.0
isodate==0.6.1
java-access-bridge-wrapper==1.2.0
jsonpath-ng==1.6.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
korean-lunar-calendar==0.3.1
lxml==5.3.0
more-itertools==10.4.0
mss==6.1.0
netsuitesdk==1.24.0
notifiers==1.3.3
numpy==2.0.1
O365==2.0.26
oauthlib==3.2.2
openpyxl==3.1.5
orderedmultidict==1.0.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pdfminer.six==20221105
pendulum==2.1.2
pillow==10.4.0
platformdirs==4.2.2
ply==3.11
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
PyMeeus==0.5.12
pynput-robocorp-fork==5.0.0
pyobjc-core==9.2
pyobjc-framework-ApplicationServices==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-Quartz==9.2
pyotp==2.9.0
pypdf==3.17.4
pyperclip==1.9.0
PySocks==1.7.1
pyspnego==0.11.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.1
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
requests_ntlm==1.3.0
robocorp-storage==1.0.4
robotframework==6.1.1
robotframework-pythonlibcore==4.4.1
robotframework-requests==0.9.7
robotframework-seleniumlibrary==6.5.0
robotframework-seleniumtestability==2.1.0
rpaframework==28.6.2
rpaframework-core==11.3.3
rpaframework-pdf==7.3.2
rpds-py==0.20.0
selenium==4.15.2
simple-salesforce==1.12.6
six==1.16.0
smartsheet-python-sdk==3.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stringcase==1.2.0
tenacity==8.5.0
trio==0.26.2
trio-websocket==0.11.1
tweepy==3.10.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==4.3.1
urllib3==1.26.15
validators==0.20.0
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zeep==4.2.1
zipp==3.20.0
